+++ addons/manager.go
@@ -2,7 +2,6 @@ package addons
 
 import (
 	"context"
-	"fmt"
 	"log/slog"
 	"os"
 
@@ -25,15 +24,12 @@ import (
 	utilruntime "k8s.io/apimachinery/pkg/util/runtime"
 	"k8s.io/client-go/kubernetes"
 	clientgoscheme "k8s.io/client-go/kubernetes/scheme"
-	"k8s.io/client-go/rest"
-	"k8s.io/client-go/tools/clientcmd"
 	"open-cluster-management.io/addon-framework/pkg/lease"
 	addonutils "open-cluster-management.io/addon-framework/pkg/utils"
 	addonapiv1alpha1 "open-cluster-management.io/api/addon/v1alpha1"
 	clusterv1 "open-cluster-management.io/api/cluster/v1"
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/cache"
-	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/manager"
 	"sigs.k8s.io/controller-runtime/pkg/metrics/server"
 )
@@ -83,6 +79,7 @@ type AddonAgentOptions struct {
 	EnableLeaderElection bool
 	ProbeAddr            string
 	HubKubeconfigFile    string
+	KubeconfigFile       string
 	SpokeClusterName     string
 	OdfOperatorNamespace string
 	DRMode               string
@@ -97,6 +94,7 @@ func (o *AddonAgentOptions) AddFlags(cmd *cobra.Command) {
 		"Enable leader election for controller manager. "+
 			"Enabling this will ensure there is only one active controller manager.")
 	flags.StringVar(&o.HubKubeconfigFile, "hub-kubeconfig", o.HubKubeconfigFile, "Location of kubeconfig file to connect to hub cluster.")
+	flags.StringVar(&o.KubeconfigFile, "kubeconfig", "", "Paths to a kubeconfig. Only required if out-of-cluster.")
 	flags.StringVar(&o.SpokeClusterName, "cluster-name", o.SpokeClusterName, "Name of spoke cluster.")
 	flags.StringVar(&o.OdfOperatorNamespace, "odf-operator-namespace", o.OdfOperatorNamespace, "Namespace of ODF operator on the spoke cluster.")
 	flags.StringVar(&o.DRMode, "mode", o.DRMode, "The DR mode of token exchange addon. Valid values are: 'sync', 'async'")
@@ -134,34 +132,10 @@ func (o *AddonAgentOptions) RunAgent(ctx context.Context) {
 	logger.Info("Addon agent has stopped")
 }
 
-// GetClientFromConfig returns a controller-runtime Client for a rest.Config
-func GetClientFromConfig(config *rest.Config) (client.Client, error) {
-	cl, err := client.New(config, client.Options{Scheme: mgrScheme})
-	if err != nil {
-		return nil, err
-	}
-	return cl, nil
-}
-
-// GetClientConfig returns the rest.Config for a kubeconfig file
-func GetClientConfig(kubeConfigFile string) (*rest.Config, error) {
-	clientconfig, err := clientcmd.BuildConfigFromFlags("", kubeConfigFile)
-	if err != nil {
-		return nil, fmt.Errorf("get clientconfig failed: %w", err)
-	}
-	return clientconfig, nil
-}
-
 func runHubManager(ctx context.Context, options AddonAgentOptions, logger *slog.Logger) {
-	hubConfig, err := GetClientConfig(options.HubKubeconfigFile)
+	hubConfig, err := utils.GetClientConfig(options.HubKubeconfigFile)
 	if err != nil {
 		logger.Error("Failed to get kubeconfig", "error", err)
-		os.Exit(1)
-	}
-
-	spokeClient, err := GetClientFromConfig(ctrl.GetConfigOrDie())
-	if err != nil {
-		logger.Error("Failed to get spoke client", "error", err)
 		os.Exit(1)
 	}
 
@@ -180,6 +154,18 @@ func runHubManager(ctx context.Context, options AddonAgentOptions, logger *slog.
 	})
 	if err != nil {
 		logger.Error("Failed to start manager", "error", err)
+		os.Exit(1)
+	}
+
+	spokeKubeConfig, err := utils.GetClientConfig(options.KubeconfigFile)
+	if err != nil {
+		logger.Error("Failed to get kubeconfig", "error", err)
+		os.Exit(1)
+	}
+
+	spokeClient, err := utils.GetClientFromConfig(spokeKubeConfig, mgr.GetScheme())
+	if err != nil {
+		logger.Error("Failed to get spoke client", "error", err)
 		os.Exit(1)
 	}
 
@@ -203,19 +189,13 @@ func runHubManager(ctx context.Context, options AddonAgentOptions, logger *slog.
 }
 
 func runSpokeManager(ctx context.Context, options AddonAgentOptions, logger *slog.Logger) {
-	hubConfig, err := GetClientConfig(options.HubKubeconfigFile)
+	spokeKubeConfig, err := utils.GetClientConfig(options.KubeconfigFile)
 	if err != nil {
 		logger.Error("Failed to get kubeconfig", "error", err)
 		os.Exit(1)
 	}
 
-	hubClient, err := GetClientFromConfig(hubConfig)
-	if err != nil {
-		logger.Error("Failed to get spoke client", "error", err)
-		os.Exit(1)
-	}
-
-	mgr, err := ctrl.NewManager(ctrl.GetConfigOrDie(), ctrl.Options{
+	mgr, err := ctrl.NewManager(spokeKubeConfig, ctrl.Options{
 		Scheme: mgrScheme,
 		Metrics: server.Options{
 			BindAddress: "0", // disable metrics
@@ -229,14 +209,13 @@ func runSpokeManager(ctx context.Context, options AddonAgentOptions, logger *slo
 		os.Exit(1)
 	}
 
-	spokeKubeConfig := mgr.GetConfig()
+	currentNamespace := os.Getenv("POD_NAMESPACE")
+
 	spokeKubeClient, err := kubernetes.NewForConfig(spokeKubeConfig)
 	if err != nil {
 		logger.Error("Failed to get spoke kube client", "error", err)
 		os.Exit(1)
 	}
-
-	currentNamespace := os.Getenv("POD_NAMESPACE")
 
 	if err = mgr.Add(manager.RunnableFunc(func(ctx context.Context) error {
 		logger.Info("Starting lease updater")
@@ -250,6 +229,18 @@ func runSpokeManager(ctx context.Context, options AddonAgentOptions, logger *slo
 		return nil
 	})); err != nil {
 		logger.Error("Failed to start lease updater", "error", err)
+		os.Exit(1)
+	}
+
+	hubConfig, err := utils.GetClientConfig(options.HubKubeconfigFile)
+	if err != nil {
+		logger.Error("Failed to get kubeconfig", "error", err)
+		os.Exit(1)
+	}
+
+	hubClient, err := utils.GetClientFromConfig(hubConfig, mgr.GetScheme())
+	if err != nil {
+		logger.Error("Failed to get hub client", "error", err)
 		os.Exit(1)
 	}
